+++ .gitignore
@@ -15,3 +15,4 @@ sonar-project.properties
 .env
 local_settings.py
 src/docker-compose-dev
+.coverage
