+++ archery/settings.py
@@ -66,6 +66,7 @@ env = environ.Env(
             "sql.notify:GenericWebhookNotifier",
         ],
     ),
+    CURRENT_AUDITOR=(str, "sql.utils.workflow_audit:AuditV2")
 )
 
 # SECURITY WARNING: keep the secret key used in production secret!
@@ -102,6 +103,8 @@ AVAILABLE_ENGINES = {
 ENABLED_NOTIFIERS = env("ENABLED_NOTIFIERS")
 
 ENABLED_ENGINES = env("ENABLED_ENGINES")
+
+CURRENT_AUDITOR = env("CURRENT_AUDITOR")
 
 # Application definition
 INSTALLED_APPS = (
